+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Sprint 5
 
+### [0.0.21] - 2025-10-20
+
+#### Changed
+
+- Take checker and come back to board
+
+#### Fixed
+
+- pylint errors
+
 ### [0.0.20] - 2025-10-18
 
 #### Changed